backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
asyncpg>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
alembic>=1.13.2
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0